+++ christmas_app/accounts.py
@@ -1,0 +1,1 @@
+""" Used to pre-defined user accounts """