+++ changelog.md
@@ -4,11 +4,14 @@
 
 # Created on December 3, 2022
 
-Config the System version in __init__.py
+Config the System version in **init**.py
 
 # Version history
+
 - Initial Create: initial created on December 3, 2022 -> **0.01**
 
 - Initial Development#1: application rules implemented on December 3, 2022 -> **0.1**
 
 - Initial Development#2: database schema implemented and initial developed on December 4, 2022 -> **0.2**
+
+- Initial Development#2.2: comprehensive application draft implemented on December 4, 2022 -> **0.22**